File for WLAN configuration of Pi Zero
In case you do not want to connect a display and a keyboard to you Raspberry Pi Zero after a fresh installation in order to configure you WLAN, you can copy this file to the boot partition of the micro SD card. The boot partition is also visible for Windows. Do not forget you modify the file with your WLAN name and password.